Move char_lit_as_u8 to its own module
